refactor(web): 認証ファクトリを auth/account の2モジュールへ分割し改名
createAccountAuthHandler を createAccountHandler へ改名し、認証クライアント操作のみの
createAuthHandler と用途で対になる名前にした(差分が Auth / Account で一目で分かる)。
単一ファイルを factory/ 配下の 3 モジュールへ分割:
- factory/context.ts: 共有の AuthHandlerContext と buildContext
- factory/auth-handler.ts: createAuthHandler
- factory/account-handler.ts: createAccountHandler

テストもファクトリ単位に分割し、実 Hono で結線する mountAuthHandler を
test/helper へ共有化した(重複を作らない)。挙動は不変(similarity 0 件・両ファクトリとも
branch カバレッジ 100%)。
