+++ application/apps/web/src/server/auth/factory/auth-handler.test.ts
@@ -4,38 +4,16 @@ import {
   UnexpectedAuthError,
 } from '@trend-diary/authentication'
 import Logger from '@trend-diary/logger'
-import { ClientError } from '@trend-diary/std/errors'
-import { type Context, Hono } from 'hono'
+import { Hono } from 'hono'
 import { HTTPException } from 'hono/http-exception'
 import { err, ok } from 'neverthrow'
 import { z } from 'zod'
 import type { Env } from '@/env'
 import CONTEXT_KEY from '@/middleware/context'
 import zodValidator from '@/middleware/zod-validator'
-import {
-  type AuthHandlerContext,
-  createAccountAuthHandler,
-  createAuthHandler,
-} from './auth-handler-factory'
-
-// resolveAccount 経路は createAccountUseCase(getRdbClient(...)) を構築するため実体依存を切る。
-// resolveAccount コールバック自身が結果を返すので use-case の中身は空で十分。
-vi.mock('@trend-diary/datastore/rdb', () => ({ default: vi.fn(() => ({})) }))
-vi.mock('@trend-diary/domain/account', () => ({ createAccountUseCase: vi.fn(() => ({})) }))
-
-// 手動 Context モック(型アサーション)を避け、実 Hono へルーティングして app.request で本番同等の
-// 経路を通す。ロガーは実 Logger(silent)を APP_LOG に載せる(handle-error.test.ts と同方針)。
-function mountHandler(handler: (c: Context<Env>) => Promise<Response>) {
-  const app = new Hono<Env>()
-  const logger = new Logger('silent')
-  app.use('*', async (c, next) => {
-    c.set(CONTEXT_KEY.APP_LOG, logger)
-    await next()
-  })
-  app.post('/auth', handler)
-  // env.DB は getRdbClient のモックが受けるため空で良い
-  return () => app.request('/auth', { method: 'POST' }, { DB: {} })
-}
+import mountAuthHandler from '@/test/helper/mount-auth-handler'
+import { createAuthHandler } from './auth-handler'
+import type { AuthHandlerContext } from './context'
 
 describe('createAuthHandler', () => {
   describe('正常系', () => {
@@ -46,7 +24,7 @@ describe('createAuthHandler', () => {
         respond: (c, output) => c.json({ count: output.items.length }, 200),
       })
 
-      const res = await mountHandler(handler)()
+      const res = await mountAuthHandler(handler)()
 
       expect(res.status).toBe(200)
       expect(await res.json()).toEqual({ count: 2 })
@@ -61,7 +39,7 @@ describe('createAuthHandler', () => {
         respond: (c) => c.json({}, 200),
       })
 
-      await mountHandler(handler)()
+      await mountAuthHandler(handler)()
 
       expect(log).toHaveBeenCalledTimes(1)
       expect(log.mock.calls[0]![0]).toEqual({ id: 'auth-1' })
@@ -74,7 +52,7 @@ describe('createAuthHandler', () => {
         respond: (c) => c.body(null, 204),
       })
 
-      const res = await mountHandler(handler)()
+      const res = await mountAuthHandler(handler)()
 
       expect(res.status).toBe(204)
       expect(await res.text()).toBe('')
@@ -124,7 +102,7 @@ describe('createAuthHandler', () => {
           respond: (c) => c.json({}, 200),
         })
 
-        const res = await mountHandler(handler)()
+        const res = await mountAuthHandler(handler)()
 
         expect(res.status).toBe(status)
       },
@@ -137,7 +115,7 @@ describe('createAuthHandler', () => {
         respond: (c) => c.json({}, 200),
       })
 
-      const res = await mountHandler(handler)()
+      const res = await mountAuthHandler(handler)()
 
       expect(res.status).toBe(403)
     })
@@ -151,72 +129,9 @@ describe('createAuthHandler', () => {
         respond: (c) => c.json({}, 200),
       })
 
-      const res = await mountHandler(handler)()
+      const res = await mountAuthHandler(handler)()
 
       expect(res.status).toBe(500)
-    })
-  })
-})
-
-describe('createAccountAuthHandler', () => {
-  describe('正常系', () => {
-    it('アカウント出力(認証出力ではない)を respond に渡すこと(log 未指定)', async () => {
-      const handler = createAccountAuthHandler({
-        createClient: () => ({}),
-        authenticate: () => Promise.resolve(ok({ id: 'auth-1' })),
-        resolveAccount: (_useCase, user) =>
-          Promise.resolve(ok({ authenticationId: user.id, displayName: 'テスト太郎' })),
-        respond: (c, account) => c.json({ displayName: account.displayName }, 200),
-      })
-
-      const res = await mountHandler(handler)()
-
-      expect(res.status).toBe(200)
-      expect(await res.json()).toEqual({ displayName: 'テスト太郎' })
-    })
-
-    it('log コールバックにアカウント出力を渡して呼ぶこと', async () => {
-      const log = vi.fn()
-      const handler = createAccountAuthHandler({
-        createClient: () => ({}),
-        authenticate: () => Promise.resolve(ok({ id: 'auth-1' })),
-        resolveAccount: () => Promise.resolve(ok({ activeUserId: 7n })),
-        log,
-        respond: (c) => c.json({}, 200),
-      })
-
-      await mountHandler(handler)()
-
-      expect(log).toHaveBeenCalledTimes(1)
-      expect(log.mock.calls[0]![0]).toEqual({ activeUserId: 7n })
-    })
-  })
-
-  describe('準正常系', () => {
-    it('認証ステップの err は toAuthError で変換されること', async () => {
-      const handler = createAccountAuthHandler({
-        createClient: () => ({}),
-        authenticate: () => Promise.resolve(err(new InvalidCredentialsError('invalid'))),
-        resolveAccount: () => Promise.resolve(ok({})),
-        respond: (c) => c.json({}, 200),
-      })
-
-      const res = await mountHandler(handler)()
-
-      expect(res.status).toBe(401)
-    })
-
-    it('アカウントステップの err は toAuthError を通さず元のステータスを保つこと', async () => {
-      const handler = createAccountAuthHandler({
-        createClient: () => ({}),
-        authenticate: () => Promise.resolve(ok({ id: 'auth-1' })),
-        resolveAccount: () => Promise.resolve(err(new ClientError('User not found', 404))),
-        respond: (c) => c.json({}, 200),
-      })
-
-      const res = await mountHandler(handler)()
-
-      expect(res.status).toBe(404)
     })
   })
 })
