+++ application/apps/web/src/server/auth/handler/logout.ts
@@ -1,5 +1,5 @@
 import { authClientConfig, PasswordAuthClient } from '@trend-diary/authentication'
-import { createAuthHandler } from '../auth-handler-factory'
+import { createAuthHandler } from '../factory/auth-handler'
 
 export default createAuthHandler({
   createClient: (c) => new PasswordAuthClient(authClientConfig(c)),
