+++ application/apps/web/src/server/auth/handler/passkey-login-verify.ts
@@ -2,7 +2,8 @@ import type { AuthenticationResponseJSON } from '@simplewebauthn/browser'
 import { authClientConfig, PasskeyClient } from '@trend-diary/authentication'
 import { z } from 'zod'
 import zodValidator from '@/middleware/zod-validator'
-import { type AuthHandlerContext, createAccountAuthHandler } from '../auth-handler-factory'
+import { createAccountHandler } from '../factory/account-handler'
+import type { AuthHandlerContext } from '../factory/context'
 
 // 真正性はSupabaseが検証するため中身の妥当性検証はプロバイダに委ね、ここは認証 ceremony 結果を素通しする
 export const passkeyLoginVerifyInputSchema = z.object({
@@ -16,7 +17,7 @@ export const passkeyLoginVerifyValidator = zodValidator('json', passkeyLoginVeri
 
 type PasskeyLoginVerifyInput = z.infer<typeof passkeyLoginVerifyInputSchema>
 
-export default createAccountAuthHandler({
+export default createAccountHandler({
   createClient: (c) => new PasskeyClient(authClientConfig(c)),
   authenticate: (client, ctx: AuthHandlerContext<PasskeyLoginVerifyInput>) =>
     client.verifyAuthentication({
