+++ application/apps/web/src/server/auth/handler/passkey-register-start.ts
@@ -1,5 +1,5 @@
 import { authClientConfig, PasskeyClient } from '@trend-diary/authentication'
-import { createAuthHandler } from '../auth-handler-factory'
+import { createAuthHandler } from '../factory/auth-handler'
 
 export default createAuthHandler({
   createClient: (c) => new PasskeyClient(authClientConfig(c)),
